refactor: separate operators tests
