+++ src/graph/tests/mod.rs
@@ -1,14 +1,6 @@
 use super::*;
-
-mod axis_op;
-mod convolution;
-mod element_wise_op;
-mod gather;
-mod max_pool;
 mod model_advanced_tests;
 mod model_tests;
-mod reduce;
+mod ops;
 mod scales_tests;
-mod softmax;
-mod typed_bin_op;
 mod utilities_tests;
