+++ src/graph/tests/ops/mod.rs
@@ -1,0 +1,8 @@
+mod axis_op;
+mod convolution;
+mod element_wise_op;
+mod gather;
+mod max_pool;
+mod reduce;
+mod softmax;
+mod typed_bin_op;
